Rename mqtt source and sink files
