+++ heisskleber/mqtt/__init__.py
@@ -1,5 +1,5 @@
 from .config import MqttConf
-from .publisher_async import AsyncMqttPublisher
-from .subscriber_async import AsyncMqttSubscriber
+from .sink import AsyncMqttPublisher
+from .source import AsyncMqttSubscriber
 
 __all__ = ["MqttConf", "AsyncMqttSubscriber", "AsyncMqttPublisher"]
